Consolidate Sass styles for posts
