--- frontend/pages/posts/python-features-in-js.module.scss
@@ -1,49 +1,0 @@
-@import '../../styles/variables.module.scss';
-
-.root {
-    flex-grow: 1;
-    background-color: $background-color;
-    overflow: hidden;
-    background-size: cover;
-    background-position: 0 400px;
-    padding-bottom: 50;
-}
-
-.link {
-    color: $link-color;
-}  
-
-.title {
-    padding: 16px;
-    max-width: 700px;
-    margin: auto;
-}
-
-.loadingState {
-    opacity: 0.05
-}
-
-.loadingMessage {
-    position: absolute;
-    top: 40%;
-    left: 40%;
-}
-
-.codeHighlight {
-    padding: 0px 48px;
-    max-width: 700px;
-    margin: auto !important;
-}
-
-.paragraph {
-    padding: 16px;
-    text-indent: 20px;
-    text-align: justify;
-    max-width: 700px;
-    margin: auto;
-}
-
-.codeSample {
-    background-color: #EDEDED;
-    font-family: Courier, monospace;
-}
